dom/l10n/tests/mochitest/document_l10n/test_docl10n_lazy.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test Lazy DocumentL10n in HTML environment</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  SimpleTest.waitForExplicitFinish();

  is(document.l10n, null, "document.l10n is null.");

  window.addEventListener("load", async () => {
    is(document.l10n, null, "document.l10n is null after load.");

    let desc = document.getElementById("main-desc");
    is(desc.textContent.length, 0, "main-desc is not translated");

    let link = document.createElement("link");
    link.setAttribute("rel", "localization");
    link.setAttribute("href", "crashreporter/aboutcrashes.ftl");
    document.head.appendChild(link);

    // Verify now that `l10n.ready` exists and is fulfilled.
    await document.l10n.ready;

    // Lazy initialized localization should translate the document.
    is(desc.textContent.length > 0, true, "main-desc is translated");

    document.head.removeChild(link);

    is(document.l10n, null, "document.l10n is null");

    SimpleTest.finish();
  }, { once: true});
  </script>
</head>
<body>
  <h1 id="main-desc" data-l10n-id="crash-reports-title"></h1>

  <p id="label1"></p>
</body>
</html>
